+++ src/main/java/com/elmakers/mine/bukkit/dao/MaterialList.java
@@ -1,7 +1,7 @@
 package com.elmakers.mine.bukkit.dao;
 
+import java.util.Collection;
 import java.util.HashSet;
-import java.util.List;
 
 import org.bukkit.Material;
 import org.bukkit.block.Block;
@@ -29,7 +29,7 @@ public class MaterialList extends HashSet<Material>
 
 	}
 
-	public MaterialList(List<Material> copy)
+	public MaterialList(Collection<Material> copy)
 	{
 		this.addAll(copy);
 	}