+++ src/main/java/com/elmakers/mine/bukkit/plugins/magic/spells/AbsorbSpell.java
@@ -1,6 +1,6 @@
 package com.elmakers.mine.bukkit.plugins.magic.spells;
 
-import java.util.List;
+import java.util.Set;
 
 import org.bukkit.Material;
 import org.bukkit.block.Block;
@@ -24,7 +24,7 @@ public class AbsorbSpell extends Spell
 			return SpellResult.NO_TARGET;
 		}
 		Material material = Material.AIR;
-		List<Material> buildingMaterials = spells.getBuildingMaterials();
+		Set<Material> buildingMaterials = spells.getBuildingMaterials();
 		byte data = 0;
 		if (!isUnderwater())
 		{